move VoFHandler implementation into single .cc
